Split az_barrio tests into 2 classes.
